fix: eliminate circular dependency
