+++ docs/v2/autoInject.js.html
@@ -80,7 +80,7 @@ import arrayMap from &apos;lodash/_arrayMap&apos;;
 import isArray from &apos;lodash/isArray&apos;;
 import trim from &apos;lodash/trim&apos;;
 import wrapAsync from &apos;./internal/wrapAsync&apos;;
-import { isAsync } from &apos;./internal/wrapAsync&apos;;
+import { isAsync } from &apos;./internal/isAsync&apos;;
 
 var FN_ARGS = /^(?:async\s+)?(function)?\s*[^\(]*\(\s*([^\)]*)\)/m;
 var FN_ARG_SPLIT = /,/;
