+++ docs/v2/ensureAsync.js.html
@@ -76,7 +76,7 @@
         <article>
             <pre class="prettyprint source linenums"><code>import setImmediate from &apos;./internal/setImmediate&apos;;
 import initialParams from &apos;./internal/initialParams&apos;;
-import { isAsync } from &apos;./internal/wrapAsync&apos;;
+import { isAsync } from &apos;./internal/isAsync&apos;;
 
 /**
  * Wrap an async function and ensure it calls its callback on a later tick of
