+++ docs/v3/asyncify.js.html
@@ -77,7 +77,7 @@
         <article>
             <pre class="prettyprint source linenums"><code>import initialParams from &apos;./internal/initialParams.js&apos;
 import setImmediate from &apos;./internal/setImmediate.js&apos;
-import { isAsync } from &apos;./internal/wrapAsync.js&apos;
+import { isAsync } from &apos;./internal/isAsync.js&apos;
 
 /**
  * Take a sync function and make it async, passing its return value to a
