+++ docs/v3/autoInject.js.html
@@ -77,7 +77,7 @@
         <article>
             <pre class="prettyprint source linenums"><code>import auto from &apos;./auto.js&apos;
 import wrapAsync from &apos;./internal/wrapAsync.js&apos;
-import { isAsync } from &apos;./internal/wrapAsync.js&apos;
+import { isAsync } from &apos;./internal/isAsync.js&apos;
 
 var FN_ARGS = /^(?:async\s+)?(?:function)?\s*\w*\s*\(\s*([^)]+)\s*\)(?:\s*{)/;
 var ARROW_FN_ARGS = /^(?:async\s+)?\(?\s*([^)=]+)\s*\)?(?:\s*=&gt;)/;
