+++ docs/v3/ensureAsync.js.html
@@ -76,7 +76,7 @@
     <section>
         <article>
             <pre class="prettyprint source linenums"><code>import setImmediate from &apos;./internal/setImmediate.js&apos;
-import { isAsync } from &apos;./internal/wrapAsync.js&apos;
+import { isAsync } from &apos;./internal/isAsync.js&apos;
 
 /**
  * Wrap an async function and ensure it calls its callback on a later tick of
