+++ lib/asyncify.js
@@ -1,6 +1,6 @@
 import initialParams from './internal/initialParams.js'
 import setImmediate from './internal/setImmediate.js'
-import { isAsync } from './internal/wrapAsync.js'
+import { isAsync } from './internal/isAsync.js'
 
 /**
  * Take a sync function and make it async, passing its return value to a
