+++ lib/autoInject.js
@@ -1,6 +1,6 @@
 import auto from './auto.js'
 import wrapAsync from './internal/wrapAsync.js'
-import { isAsync } from './internal/wrapAsync.js'
+import { isAsync } from './internal/isAsync.js'
 
 var FN_ARGS = /^(?:async\s+)?(?:function)?\s*\w*\s*\(\s*([^)]+)\s*\)(?:\s*{)/;
 var ARROW_FN_ARGS = /^(?:async\s+)?\(?\s*([^)=]+)\s*\)?(?:\s*=>)/;
