+++ lib/ensureAsync.js
@@ -1,5 +1,5 @@
 import setImmediate from './internal/setImmediate.js'
-import { isAsync } from './internal/wrapAsync.js'
+import { isAsync } from './internal/isAsync.js'
 
 /**
  * Wrap an async function and ensure it calls its callback on a later tick of
