+++ lib/internal/eachOfLimit.js
@@ -1,7 +1,7 @@
 import once from './once.js'
 import iterator from './iterator.js'
 import onlyOnce from './onlyOnce.js'
-import {isAsyncGenerator, isAsyncIterable} from './wrapAsync.js'
+import { isAsyncGenerator, isAsyncIterable } from './isAsync.js'
 import asyncEachOfLimit from './asyncEachOfLimit.js'
 import breakLoop from './breakLoop.js'
