+++ lib/internal/wrapAsync.js
@@ -1,16 +1,5 @@
 import asyncify from '../asyncify.js'
-
-function isAsync(fn) {
-    return fn[Symbol.toStringTag] === 'AsyncFunction';
-}
-
-function isAsyncGenerator(fn) {
-    return fn[Symbol.toStringTag] === 'AsyncGenerator';
-}
-
-function isAsyncIterable(obj) {
-    return typeof obj[Symbol.asyncIterator] === 'function';
-}
+import { isAsync } from './isAsync.js'
 
 function wrapAsync(asyncFn) {
     if (typeof asyncFn !== 'function') throw new Error('expected a function')
@@ -18,5 +7,3 @@ function wrapAsync(asyncFn) {
 }
 
 export default wrapAsync;
-
-export { isAsync, isAsyncGenerator, isAsyncIterable };
