+++ lib/retryable.js
@@ -1,6 +1,7 @@
 import retry from './retry.js'
 import initialParams from './internal/initialParams.js'
-import {default as wrapAsync, isAsync} from './internal/wrapAsync.js'
+import wrapAsync from './internal/wrapAsync.js'
+import { isAsync } from './internal/isAsync.js'
 import { promiseCallback, PROMISE_SYMBOL } from './internal/promiseCallback.js'
 
 /**
